Intern 75

See merge request !33

+++ src/AppBundle/Admin/NewsAdmin.php
@@ -7,7 +7,7 @@ use Sonata\AdminBundle\Admin\AbstractAdmin;
 use Sonata\AdminBundle\Datagrid\ListMapper;
 use Sonata\AdminBundle\Form\FormMapper;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
-
+use Cocur\Slugify\Slugify;
 
 class NewsAdmin extends AbstractAdmin
 {
@@ -21,8 +21,13 @@ class NewsAdmin extends AbstractAdmin
         $dbservice = $this->getConfigurationPool()->getContainer()->get('app.database_news');
         $resQuery = $dbservice->getNewsBySlug($slug);
 
-        $imageName = explode('/', $resQuery[0]->getImage());
-        $imageName = $imageName[count($imageName)-1];
+        if (!empty($slug)) {
+            $imageName = explode('/', $resQuery[0]->getImage());
+            $imageName = $imageName[count($imageName)-1];
+        }
+        else {
+            $imageName = '';
+        }
 
         return $imageName;
     }
@@ -47,7 +52,8 @@ class NewsAdmin extends AbstractAdmin
                 'label' => 'Содержание'
             ])
             ->add('active', 'checkbox', [
-                'label' => 'Активен'
+                'label' => 'Активен',
+                'required' => false
             ])
             ->add('description', 'text', [
                 'label' => 'Описание'
@@ -87,10 +93,9 @@ class NewsAdmin extends AbstractAdmin
             ->addIdentifier('image')
             ->addIdentifier('slug')
             ->addIdentifier('publicationDate', 'datetime', [
-                'format' => 'Y-m-d',
+                'format' => 'd.m.Y',
                 'timezone' => 'America/New_York'
             ])
-            ->addIdentifier('content')
             ->addIdentifier('active')
             ->addIdentifier('description')
             ->add('_action', 'actions', [
@@ -105,7 +110,11 @@ class NewsAdmin extends AbstractAdmin
     {
         $path = $this->getConfigurationPool()->getContainer()->getParameter('img_path');
         $imgService = $this->getConfigurationPool()->getContainer()->get('app.image_manager');
+        $cocur = $this->getConfigurationPool()->getContainer()->get('cocur_slugify');
         $correctPath = $imgService->upload($news, $path);
+
+        $slug = $cocur->activateRuleset('russian')->slugify($news->getTitle());
+        $news->setSlug($slug);
 
         if ($news->getDel() == 1 && !empty($news->getImage())) {
             $imgService->remove($news->getImage());
@@ -115,11 +124,10 @@ class NewsAdmin extends AbstractAdmin
             if(!empty($correctPath) && empty($news->getImage())){
                 $news->setImage($correctPath);
             }
-            else {
-                if ($correctPath != $news->getImage()) {
-                    $imgService->remove($news->getImage());
-                    $news->setImage($correctPath);
-                }
+            elseif ($correctPath != $news->getImage() && $news->getActive() == 1) {
+                $imgService->remove($news->getImage());
+                $news->setImage($correctPath);
+
             }
         }
